Support for symlinks. Rename FileSystem to FileManager for Path class.
