+++ TestRequirements/Fixtures/ProjectWithTests/PublicDirectory/Symlink
@@ -1,0 +1,1 @@
+../../SymlinkSource
