+++ Tests/FileManager/Directory/DeleteRecursiveTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\FileManager\Directory\DeleteRecursiveTest;
 
-use Saeghe\Saeghe\FileSystem\Address;
+use Saeghe\Saeghe\FileManager\Address;
 use function Saeghe\Saeghe\FileManager\Directory\delete_recursive;
 
 test(
